backend/requirements.txt
django
mysqlclient
python-dotenv
djangorestframework
django-cors-headers
google-generativeai
google-genai
firebase-admin